+++ Split_Image_Explorer.py
@@ -1,15 +1,16 @@
 from utils import *
 from auto_rotate_geotiff import *
-
-from PyQt5.QtWidgets import QWidget, QApplication, QHBoxLayout, QLabel, QVBoxLayout, QCheckBox, QSlider, QLineEdit, QPushButton, QButtonGroup
+#CST20240312
+from PyQt5.QtWidgets import QWidget, QApplication, QLabel, QHBoxLayout, QVBoxLayout, QCheckBox, QSlider, QLineEdit, QPushButton, QButtonGroup, QMainWindow, QGridLayout
 #CST 20240308
-
-from PyQt5.QtGui import QFont, QPixmap
-from PyQt6.QtGui import QImage, QGuiApplication
+from PyQt5.QtGui import QPixmap, QImage, QFont, QGuiApplication, QFont, qRgb
 
 from PyQt5.QtCore import Qt
 
 from PIL.ImageQt import ImageQt
+import os
+import shutil
+
 
 from Models import *
 from Dataset import *
@@ -25,18 +26,23 @@ import yaml
 
 
 class SplitImageTool(QWidget):
-
+    global numTiff
+    numTiff = 0
     def __init__(self, cfg_path, checkpoint=None, netcdf=False):
         super().__init__()
 
         # Initialize GUI Window properties
         print('-------- Initializing App --------')
+        
+        # geometry
         screen_resolution = app.desktop().availableGeometry()
         self.title = 'Split Image Labeling tool'
         self.width, self.height = int(screen_resolution.width()), int(screen_resolution.height())
-        self.setGeometry(0, 0, self.width, self.height)
+        self.setMinimumSize(self.width - 100, self.height - 100) # the default min size runs of the screen!
+        self.setGeometry(0, 0, self.width - 100, self.height - 100)
         self.setWindowTitle(self.title)
         self.to_netcdf = netcdf
+
 
         # Load Tiff Image and split image data
         print('-------- Loading App Config --------')
@@ -50,6 +56,7 @@ class SplitImageTool(QWidget):
         self.tiff_selector = 0
         self.checkpoint = checkpoint
         self.label_path = self.cfg['npy_path']
+        self.lagstep = self.cfg['vario_num_lag']
         self.label_data = np.load(self.label_path, allow_pickle=True)
         self.split_info_save = self.label_data[1]
 
@@ -96,6 +103,8 @@ class SplitImageTool(QWidget):
         self.setLayout(self.master_layout)
         self.show()
 
+        
+
     def initDataset(self):
         self.dataset_info = self.label_data[0]
         self.split_info = self.split_info_save[self.split_info_save[:,6] == self.tiff_selector]
@@ -107,6 +116,8 @@ class SplitImageTool(QWidget):
         # Load classification results if specified
         if self.checkpoint != None:
             self.pred_label_path = self.checkpoint
+            #CST 20240313
+            print("label path", self.pred_label_path)
             pred_data = np.load(self.pred_label_path, allow_pickle=True)
             self.pred_labels_save = pred_data[1]
             self.predictions = True
@@ -151,6 +162,7 @@ class SplitImageTool(QWidget):
         self.visualization_save_buttons = QVBoxLayout()
         self.split_text = QHBoxLayout()
         self.class_buttons = QVBoxLayout()
+        self.grid_layout = QGridLayout()
 
         # --- Initialize UI elements ---
         # Current split Image container
@@ -158,6 +170,11 @@ class SplitImageTool(QWidget):
         self.split_image_label = QLabel(self)
         self.split_image_label.setMargin(0)
         self.split_image_label.setAlignment(Qt.AlignCenter)
+        #self.split_image_labels = [QLabel() for _ in self.lagstep]
+        self.split_image_labels = [QLabel(self) for _ in range(4)]
+        for i in range(4):
+            self.split_image_labels[i].setMargin(0)
+            self.split_image_labels[i].setAlignment(Qt.AlignCenter)
 
         # Background Image container
         self.tiff_image_pixmap = QPixmap()
@@ -213,6 +230,10 @@ class SplitImageTool(QWidget):
         self.save_heatmap_button = QPushButton('Save Heatmap Image')
         self.save_heatmap_button.clicked.connect(self.saveHeatmapCallback)
 
+         #SAVE CONFIDENCE PREDICTIONS
+        self.save_confidente_predictions_button = QPushButton('Save Confidence Predictions')
+        self.save_confidente_predictions_button.clicked.connect(self.savePredictionsCallbackNPY)
+
         if self.has_contour == False:
             self.save_contour_button = QPushButton('Save Contour File')
             self.save_contour_button.clicked.connect(self.saveContourCallback)
@@ -222,7 +243,6 @@ class SplitImageTool(QWidget):
         self.conf_thresh_value.setMargin(0)
         self.conf_thresh_value.setAlignment(Qt.AlignCenter)
         self.conf_thresh_value.setFont(QFont("Helvetica", 13, QFont.Bold))
-
         self.conf_thresh_slider = QSlider(Qt.Horizontal)
         self.conf_thresh_slider.setFocusPolicy(Qt.StrongFocus)
         self.conf_thresh_slider.setTickPosition(QSlider.TicksBothSides)
@@ -231,6 +251,7 @@ class SplitImageTool(QWidget):
         self.conf_thresh_slider.setMinimum(0)
         self.conf_thresh_slider.setMaximum(99)
         self.conf_thresh_slider.setValue(0)
+        
         low = np.floor(np.array(self.conf_cmap.colors[0])*255)
         mid = np.floor(np.array(self.conf_cmap.colors[50])*255)
         high = np.floor(np.array(self.conf_cmap.colors[99])*255)
@@ -245,46 +266,71 @@ class SplitImageTool(QWidget):
         self.split_text.addWidget(self.split_image_conf)
         self.split_text.addSpacing(200)
 
-        self.visualization_widgets.addWidget(self.split_image_label)
+        # self.visualization_widgets.addWidget(self.split_image_label)
+        for image_label in self.split_image_labels:
+            self.visualization_widgets.addWidget(image_label)
         self.visualization_widgets.addLayout(self.split_text)
+        self.visualization_widgets.addWidget(self.split_image_label)
+        
 
         self.conf_slider_container.addWidget(self.conf_thresh_slider)
         self.conf_slider_container.addWidget(self.conf_thresh_value)
         self.visualization_widgets.addLayout(self.conf_slider_container)
+        
 
         self.visualization_toggles.addWidget(self.predictions_button)
         self.visualization_toggles.addWidget(self.heatmap_button)
         self.visualization_toggles.addWidget(self.labels_button)
 
         self.visualization_save_buttons.addWidget(self.save_predictions_button)
+        #INITIALIZE THE BUTTON THAT SAVES THE CONFIDENT PREDICTIONS 
+        self.visualization_save_buttons.addWidget(self.save_confidente_predictions_button)
         self.visualization_save_buttons.addWidget(self.save_heatmap_button)
         if self.has_contour == False:
             self.visualization_save_buttons.addWidget(self.save_contour_button)
 
         self.visualization_interactive.addLayout(self.visualization_toggles)
+        
         self.visualization_interactive.addLayout(self.visualization_save_buttons)
+        
 
         self.visualization_widgets.addLayout(self.visualization_interactive)
+        
 
         self.new_class_layout = QHBoxLayout()
         self.new_class_layout.addSpacing(200)
         self.new_class_layout.addWidget(self.new_class_field)
         self.new_class_layout.addSpacing(200)
         self.visualization_widgets.addLayout(self.new_class_layout)
+        
 
-        self.tiff_selector_buttons = QHBoxLayout()
+        self.tiff_selector_buttons = QGridLayout()
 
         for tiffNum in range(len(self.dataset_info['filename'])):
             button = QPushButton('{}...'.format(self.dataset_info['filename'][tiffNum].split('/')[-1][:13]), self)
             button.clicked.connect(self.makeTiffSelectorCallbacks(tiffNum))
-            self.tiff_selector_buttons.addWidget(button)
+            
+            # Calculate the row and column numbers
+            row = tiffNum // 4
+            col = tiffNum % 4
+            
+            # Add the button to the layout at the specified row and column
+            self.tiff_selector_buttons.addWidget(button, row, col)
 
         self.left_layout.addLayout(self.tiff_selector_buttons)
+        
         self.left_layout.addLayout(self.visualization_widgets)
+        
         self.left_layout.addLayout(self.class_buttons)
-
+        
+        self.left_layout.addLayout(self.grid_layout) #something about this causes a warning: QLayout::addChildLayout: layout "" already has a parent
+        # this might be the issue? as in the layout is added to the master too early?
         self.master_layout.addLayout(self.left_layout)
+        
         self.master_layout.addWidget(self.tiff_image_label)
+
+        
+
 
     def addClassButton(self, i, className, container):
         buttonContainer = QHBoxLayout()
@@ -305,18 +351,19 @@ class SplitImageTool(QWidget):
         buttonContainer.addWidget(labelButton)
         buttonContainer.addSpacing(20)
         container.addLayout(buttonContainer)
+        
 
     def initClassButtons(self):
 
-        numColumns = math.ceil(len(self.class_enum) / 12)
+        numColumns = math.ceil(len(self.class_enum) / 4)
 
         self.class_buttons_columns_list = []
         for i in range(numColumns):
             self.class_buttons_columns_list.append(QVBoxLayout())
 
         for i, className in enumerate(self.class_enum):
-            col = math.floor(i / 12)
-            row = i % 12
+            col = math.floor(i / 4)
+            row = i % 4
             self.addClassButton(i, className, self.class_buttons_columns_list[col])
 
         self.class_buttons_columns = QHBoxLayout()
@@ -324,9 +371,8 @@ class SplitImageTool(QWidget):
             self.class_buttons_columns.addLayout(column)
 
         self.class_buttons.addLayout(self.class_buttons_columns)
-
+    
     def initBgImage(self):
-
         # Scale down tiff image for visualization and convert to 8-bit RGB
         scale_factor = int(self.tiff_image_matrix.shape[0] / 1200)
         bg_img_scaled = self.tiff_image_matrix[::scale_factor,::scale_factor]
@@ -357,13 +403,11 @@ class SplitImageTool(QWidget):
 
         # Convert to QImage from cv and wrap in QPixmap container
         #CST 20240312
-        self.bg_qimg = QImage(self.bg_img_cv.data,self.bg_img_cv.shape[1],self.bg_img_cv.shape[0],self.bg_img_cv.shape[1]*3,QImage.Format.Format_RGB888)
+        self.bg_qimg = QImage(self.bg_img_cv.data,self.bg_img_cv.shape[1],self.bg_img_cv.shape[0],self.bg_img_cv.shape[1]*3,QImage.Format_RGB888)
         self.tiff_image_pixmap = QPixmap(self.bg_qimg)
-        self.tiff_image_pixmap = QPixmap(self.tiff_image_pixmap.scaledToWidth(int(self.width/2) - 10))
+        self.tiff_image_pixmap = self.tiff_image_pixmap.scaledToWidth(int(self.width/2) - 10)
         #self.tiff_image_pixmap = self.tiff_image_pixmap.scaledToHeight(int(self.height - 50)).scaledToWidth(int(self.width/2) - 10)
         #print('bg_img_pixmap:  {}x{}'.format(self.tiff_image_pixmap.size().height(), self.tiff_image_pixmap.size().width()))
-        #CST 20240312
-        print(type(self.tiff_image_pixmap))
         # Get scaling factor between cv and q image
         bg_img_cv_size = np.array(self.bg_img_cv.shape[:-1])
         bg_img_q_size = np.array((self.tiff_image_pixmap.size().height(), self.tiff_image_pixmap.size().width()))
@@ -373,63 +417,148 @@ class SplitImageTool(QWidget):
 
     def updateBgImage(self):
         return
+    #For graphing variograms
+    def draw_line(self, img, x1, y1, x2, y2, color):
+        """Draw a simple line on a QImage."""
+        dx = abs(x2 - x1)
+        dy = abs(y2 - y1)
+        sx = 1 if x1 < x2 else -1
+        sy = 1 if y1 < y2 else -1
+        err = dx - dy
+
+        while True:
+            if 0 <= x1 < img.width() and 0 <= y1 < img.height():
+                img.setPixel(x1, y1, color)
+            if x1 == x2 and y1 == y2:
+                break
+            e2 = 2 * err
+            if e2 > -dy:
+                err -= dy
+                x1 += sx
+            if e2 < dx:
+                err += dx
+                y1 += sy
+
+    def draw_variogram_on_image(self, vario, qimage):
+        width = qimage.width()
+        height = qimage.height()
+        
+        # Normalize vario values to fit within image dimensions
+        max_value = max(vario)
+        min_value = min(vario)
+        
+        for i in range(len(vario) - 1):
+            x1 = int(i * width / len(vario))
+            y1 = int((vario[i] - min_value) * height / (max_value - min_value))
+            x2 = int((i + 1) * width / len(vario))
+            y2 = int((vario[i + 1] - min_value) * height / (max_value - min_value))
+            
+            # Ensure coordinates are within bounds
+            y1 = height - min(max(y1, 0), height - 1)
+            y2 = height - min(max(y2, 0), height - 1)
+            
+            # Draw line on the image (basic line drawing algorithm)
+            self.draw_line(qimage, x1, y1, x2, y2, qRgb(0, 0, 0))
 
     def getNewImage(self, index):
-
-         self.image_index = index
+        
+        self.image_index = index
 
          # Grab info of split image at index
-         x,y,x_utm,y_utm,label,conf,_ = self.split_info[index]
-         x,y,x_utm,y_utm,label = int(x),int(y),int(x_utm),int(y_utm),int(label)
+        x,y,x_utm,y_utm,label,conf,_ = self.split_info[index]
+        x,y,x_utm,y_utm,label = int(x),int(y),int(x_utm),int(y_utm),int(label)
+        #CST20240313
+        # Get split image from image matrix
+        img = self.tiff_image_matrix[x:x+self.win_size[0],y:y+self.win_size[1]]
+        img = scaleImage(img, self.tiff_image_max)
+        variograms = silas_directional_vario(img)
+        qimg = QImage(img.data, img.shape[1], img.shape[0], img.shape[1], QImage.Format_Grayscale8)
+        image = Image.fromarray(img).convert("L")
+         # Wrap split image in QPixmap       
+        self.split_image_pixmap = QPixmap.fromImage(qimg).scaledToWidth(270)
+        self.split_image_label.setPixmap(self.split_image_pixmap)
+        
 
-         # Get split image from image matrix
-         img = self.tiff_image_matrix[x:x+self.win_size[0],y:y+self.win_size[1]]
-         img = scaleImage(img, self.tiff_image_max)
-         img = Image.fromarray(img).convert("L")
-         img = ImageQt(img)
+        self.variograms = variograms
+        
+        
+        for i, vario in enumerate(self.variograms):
+            qimage = QImage(img.shape[1], img.shape[0], QImage.Format_RGB32)
+            qimage.fill(Qt.white)
 
-         # Wrap split image in QPixmap
-         
-         self.split_image_pixmap = QPixmap.fromImage(img).scaledToWidth(270)
-         self.split_image_label.setPixmap(self.split_image_pixmap)
-         # Update label text
-         class_text = ''
-         if self.predictions:
+            # Convert variogram data to image plot
+            self.draw_variogram_on_image(vario, qimage)
+            
+            row = i // 2
+            col = i % 2
+
+            # Wrap split image in QPixmap
+            split_image_pixmap = QPixmap.fromImage(qimage).scaledToWidth(270)
+
+            # Set the pixmap of the QLabel at the current index
+            self.split_image_labels[i].setPixmap(split_image_pixmap)
+            self.grid_layout.addWidget(self.split_image_labels[i], row, col)
+        # Update label text
+        class_text = ''
+        if self.predictions:
             label = int(self.pred_labels[index,4])
             conf = self.pred_labels[index,5]
             class_text = "Class %d: %s"%(label, self.class_enum[label])
             class_conf = "Conf: {:.2%}".format(conf)
-         elif label == -1:
+        elif label == -1:
             class_text = "No class assigned yet"
             class_conf = ''
-         else:
+        else:
             class_text = "Class %d: %s"%(label, self.class_enum[label])
             class_conf = "Conf: {:.2%}".format(conf)
 
-         self.split_image_class.setText(class_text)
-         self.split_image_conf.setText(class_conf)
+        self.split_image_class.setText(class_text)
+        self.split_image_conf.setText(class_conf)
 
-         bg_img = self.bg_img_cv.copy()
-         height,width,_ = self.bg_img_cv.shape
-         imgSize = np.array([width,height])
+        bg_img = self.bg_img_cv.copy()
+        height,width,_ = self.bg_img_cv.shape
+        imgSize = np.array([width,height])
 
-         pix_coords = utm_to_pix(imgSize, self.bg_img_utm.T, np.array([[x_utm,y_utm]]))
+        pix_coords = utm_to_pix(imgSize, self.bg_img_utm.T, np.array([[x_utm,y_utm]]))
 
-         # Draw crosshairs
-         cv2.circle(bg_img,(pix_coords[0][0],height-pix_coords[0][1]),4,(255,0,0),thickness=-1)
-         cv2.line(bg_img, (pix_coords[0][0], 0), (pix_coords[0][0], height), (255,0,0),thickness=2)
-         cv2.line(bg_img, (0, height-pix_coords[0][1]), (width, height-pix_coords[0][1]), (255,0,0),thickness=2)
+        # Draw crosshairs
+        cv2.circle(bg_img,(pix_coords[0][0],height-pix_coords[0][1]),4,(255,0,0),thickness=-1)
+        cv2.line(bg_img, (pix_coords[0][0], 0), (pix_coords[0][0], height), (255,0,0),thickness=2)
+        cv2.line(bg_img, (0, height-pix_coords[0][1]), (width, height-pix_coords[0][1]), (255,0,0),thickness=2)
 
-         height,width,channels = bg_img.shape
-         #CST20240312
-         bg_img = QImage(bg_img.data,width,height,width*channels,QImage.Format.Format_RGB888)
-         background_image = QPixmap(bg_img)
-         #background_image = background_image.scaledToHeight(int(self.height - 50)).scaledToWidth(int(self.width/2) - 10)
-         background_image = background_image.scaledToWidth(int(self.width/2) - 10)
+        height,width,channels = bg_img.shape
+        #CST20240312
+        bg_img = QImage(bg_img.data,width,height,width*channels,QImage.Format_RGB888)
+        background_image = QPixmap(bg_img)
+        #background_image = background_image.scaledToHeight(int(self.height - 50)).scaledToWidth(int(self.width/2) - 10)
+        background_image = background_image.scaledToWidth(int(self.width/2) - 10)
 
-         self.tiff_image_label.setPixmap(background_image)
+        self.tiff_image_label.setPixmap(background_image)
 
+    #CST20240313 (creating a function to write images into a folder)
+    def writeImage(self,filePath, fileName, index):
+        p = os.path.join(filePath, fileName)
+        fp = p + '.png'
+        self.image_index = index
+         # Grab info of split image at index
+        x,y,x_utm,y_utm,label,conf,_ = self.split_info[index]
+        x,y,x_utm,y_utm,label = int(x),int(y),int(x_utm),int(y_utm),int(label)
+         # Get split image from image matrix
+        img = self.tiff_image_matrix[x:x+self.win_size[0],y:y+self.win_size[1]]
+        img = scaleImage(img, self.tiff_image_max)
+        image = QImage(img.data, img.shape[1], img.shape[0], img.shape[1], QImage.Format_Grayscale8)
+        image.save(fp,"png")
+        
 
+ #CST20240403 Checks all directories for the image, and deletes it if it finds the image.
+    def deleteImage(self,filePath,fileName):
+        numClasses = cfg['num_classes']
+        for i in range(numClasses):
+            file_path = (filePath+str(i)+'/'+str(i)+fileName+'.png')
+            if os.path.isfile(file_path):
+                os.remove(file_path)
+                
+    
     def label(self, mask, class_label):
         self.split_info[mask,4] = class_label
         self.split_info[mask,5] = 1
@@ -439,7 +568,20 @@ class SplitImageTool(QWidget):
     def labelCurrent(self, class_label):
         self.split_info[self.image_index][4] = class_label
         self.split_info[self.image_index][5] = 1
-        self.getNewImage(self.image_index)
+                #Load training img path
+        if self.cfg['training_img_path'] != 'None':
+            labeled_img_path = cfg['training_img_path']
+            if not os.path.exists(labeled_img_path+"/"): os.mkdir(labeled_img_path+"/")
+            if not os.path.exists(labeled_img_path+"/"+str(class_label)): os.mkdir(labeled_img_path+"/"+str(class_label))
+            self.deleteImage(labeled_img_path+"/", str(self.image_index)+str(numTiff))
+            self.writeImage(labeled_img_path+"/"+str(class_label), str(class_label)+str(self.image_index)+str(numTiff), self.image_index)
+            self.getNewImage(self.image_index)
+        else:
+            if not os.path.exists("Classifications/"): os.mkdir("Classifications/")
+            if not os.path.exists("Classifications/"+str(class_label)): os.mkdir("Classifications/"+str(class_label))
+            self.deleteImage("Classifications/", str(self.image_index)+str(numTiff))
+            self.writeImage("Classifications/"+str(class_label), str(class_label)+str(self.image_index)+str(numTiff), self.image_index)
+            cfg['training_img_path'] = 'Classifications'
         self.update()
 
     def batchSelectLabel(self, class_label):
@@ -451,9 +593,22 @@ class SplitImageTool(QWidget):
             if Point(img[2],img[3]).within(batch_select):
                 self.split_info[i][4] = class_label
                 self.split_info[i][5] = 1
+                if self.cfg['training_img_path'] != 'None':
+                    labeled_img_path = self.cfg['training_img_path']
+                    if not os.path.exists(labeled_img_path+"/"): os.mkdir(labeled_img_path+"/")
+                    if not os.path.exists(labeled_img_path+"/"+str(class_label)): os.mkdir(labeled_img_path+"/"+str(class_label))
+                    self.deleteImage(labeled_img_path+"/", str(i)+str(numTiff))
+                    self.writeImage(labeled_img_path+"/"+str(class_label), str(class_label)+str(i)+str(numTiff), i)
+                else:
+                    if not os.path.exists("Classifications/"): os.mkdir("Classifications/")
+                    if not os.path.exists("Classifications/"+str(class_label)): os.mkdir("Classifications/"+str(class_label))
+                    self.deleteImage("Classifications/", str(i)+str(numTiff))
+                    self.writeImage("Classifications/"+str(class_label), str(class_label)+str(i)+str(numTiff), i)
+                    cfg['training_img_path'] = 'Classifications'
         self.batch_select_polygon = []
         self.getNewImage(self.image_index)
         self.update()
+
 
     def getMousePosUTM(self, event):
 
@@ -513,7 +668,7 @@ class SplitImageTool(QWidget):
 
             height,width,channels = bg_img.shape
             #CST 202040312
-            bg_img = QImage(bg_img.data,width,height,width*channels,QImage.Format.Format_RGB888)
+            bg_img = QImage(bg_img.data,width,height,width*channels,QImage.Format_RGB888)
             background_image = QPixmap(bg_img)
             #background_image = background_image.scaledToHeight(int(self.height - 50)).scaledToWidth(int(self.width/2) - 10)
             background_image = background_image.scaledToWidth(int(self.width/2) - 10)
@@ -596,7 +751,12 @@ class SplitImageTool(QWidget):
             self.initBgImage()
             self.getNewImage(0)
         return tiff_selector_callback
-
+    def getTiffnum(self, tiff):
+        def getNum():
+            global numTiff
+            numTiff = tiff
+            #print("GEOTIFF number", numTiff)
+        return getNum
     @pyqtSlot()
     def on_click(self):
         self.update()
@@ -661,7 +821,7 @@ class SplitImageTool(QWidget):
         #CST20240308
         else:
             out_path = self.pred_label_path[:-4] + '_prediction.png'
-            print("savePredictionsCallback", out_path)
+            ("Prediction image saved to", out_path)
 
 
     def saveHeatmapCallback(self):
@@ -673,7 +833,95 @@ class SplitImageTool(QWidget):
             #CST20240308
         else:
             out_path = self.pred_label_path[:-4] + '_confidence_heatmap.png'
-            print("saveHeatmapCallback", out_path)
+            print("Heat map saved to", out_path)
+
+    #function that saves the confident predictions
+    def savePredictionsCallbackNPY(self):
+        savepred = cfg['save_all_pred']
+        saveMin = cfg['equal_dataset']
+    
+        #check if there are predictions loaded so that app doesn't crash
+        if self.checkpoint == None:
+            print('No predictions loaded')
+            return
+        #create the directory if it does not exist
+        dirName = 'ConfidentPredictions'
+        if not os.path.exists(dirName):
+            os.mkdir(dirName)
+        #create the filename
+        filename = f'ConfidentPredictions/confidence_predictions_{self.conf_thresh}.npy'
+
+        #get the dataset
+        dataset_path = cfg['npy_path']   
+        dataset = np.load(dataset_path, allow_pickle=True)
+        if saveMin == False:
+            if savepred == False:
+                # Save predicitions above the confidence threshold
+                self.confident_predictions = self.pred_labels[self.pred_labels[:,5] > self.conf_thresh]        
+
+                dataset[1] = self.confident_predictions #update the dataset with the new confident predictions
+                
+            else: #Should save all WV datasets, not just the one selected.
+                self.confident_predictions = self.pred_labels_save[self.pred_labels_save[:,5] > self.conf_thresh]
+                dataset[1] = self.confident_predictions
+            np.save(filename, dataset) #save the dataset as npy file
+            print('File saved to', filename)
+        else:
+            total = 0
+            numClasses = cfg['num_classes']
+            minSize =  100000
+            classSize = 0
+            classes = 0
+            predictions = []
+            if savepred == False:
+                
+                # Save predicitions above the confidence threshold
+                for i in range(numClasses):
+                    data = self.pred_labels[self.pred_labels[:,4] == i]
+                    self.confident_predictions = data[data[:,5] > self.conf_thresh]
+                    classSize = len(self.confident_predictions)
+                    if classSize < minSize: 
+                        minSize = classSize
+                for i in range(numClasses): 
+                    data = self.pred_labels[self.pred_labels[:,4] == i]
+                    self.confident_predictions = data[data[:,5] > self.conf_thresh]
+                    if classSize != 0:
+                        for i in range(minSize): #Should select the highest confidence images from each class
+                                highest_confidence_index = np.argmax(self.confident_predictions[:, 5])
+                                predictions.append(self.confident_predictions[highest_confidence_index])
+                                self.confident_predictions = np.delete(self.confident_predictions, highest_confidence_index, axis=0)
+                                total += 1
+                predictions = np.array(predictions)
+                dataset[1] = predictions #update the dataset with the new confident predictions
+                
+            else: #Should save all WV datasets, not just the one selected.
+                # Save predicitions above the confidence threshold
+                for i in range(numClasses):
+                    data = self.pred_labels_save[self.pred_labels_save[:,4] == i]
+                    self.confident_predictions = data[data[:,5] > self.conf_thresh]
+                    classSize = len(self.confident_predictions)
+                    if classSize != 0:
+                        print("Saving images from class ", i)
+                        print("Class: ", i, "Number of Images: ", classSize)
+                        classes += 1
+                        if classSize < minSize: 
+                            minSize = classSize                           
+                for i in range(numClasses): 
+                    data = self.pred_labels_save[self.pred_labels_save[:,4] == i]
+                    self.confident_predictions = data[data[:,5] > self.conf_thresh]
+                    classSize = len(self.confident_predictions)
+                    if classSize != 0:
+                        for i in range(minSize):  #Should select the highest confidence images from each class
+                            highest_confidence_index = np.argmax(self.confident_predictions[:, 5])
+                            predictions.append(self.confident_predictions[highest_confidence_index])
+                            self.confident_predictions = np.delete(self.confident_predictions, highest_confidence_index, axis=0)
+                            total += 1
+                predictions = np.array(predictions)
+                dataset[1] = predictions #update the dataset with the new confident predictions
+            print(minSize, "Images saved in for each class for a total dataset size of ", total, "images from ", classes, "crevasse classes")
+            np.save(filename, dataset) #save the dataset as npy file
+            print('File saved to', filename)
+
 
     def saveContourCallback(self):
         if self.batch_select_polygon != []:
@@ -701,7 +949,7 @@ class SplitImageTool(QWidget):
         self.cfg['class_enum'] = self.class_enum
         self.cfg['num_classes'] = len(self.class_enum)
         f = open(args.config, 'w')
-        f.write(generate_config(self.cfg))
+        f.write(generate_config_silas(self.cfg))
         f.close()
 
         if self.to_netcdf:
@@ -712,7 +960,6 @@ class SplitImageTool(QWidget):
         return
 
 if __name__ == '__main__':
-
     # Parse command line flags
     
     parser = argparse.ArgumentParser()
@@ -720,9 +967,12 @@ if __name__ == '__main__':
     parser.add_argument("--load_labels", type=str, default=None)
     parser.add_argument("--netcdf", action="store_true")
     args = parser.parse_args()
+    # Read config file
+    with open(args.config, 'r') as ymlfile:
+        cfg = yaml.load(ymlfile, Loader=yaml.FullLoader)
 
     app = QApplication(sys.argv)
-    QGuiApp = QGuiApplication(sys.argv)
+    QGuiApp = QApplication(sys.argv)
     ex = SplitImageTool(args.config, args.load_labels, args.netcdf)
 
     sys.exit(app.exec_())
